UI 1/UI/selectors/learned_selector_3/pyproject.toml
[project]
name = "Learned Selector - Induct, Yard, Usage"
version = "0.1.0"
description = "My demo selector"
authors = [{ name = "John Doe", email = "john.doe@composabl.com" }]
dependencies = [
    "composabl-core"
]

[composabl]
type = "selector-teacher"
entrypoint = "learned_selector_3.teacher:UPSTeacher"
